# api-doc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job API",
        "version": "0.1"
    },
    "paths": {
...
        "/route": {
            "get": {
                "tags": [
                    "Route"
                ],
                "description": "Returns a list of route objects",
                "operationId": "Abc\\Job\\Controller\\RouteController::all",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Route"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "In case of an internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Route"
                ],
                "description": "Creates a route",
                "operationId": "Abc\\Job\\Controller\\RouteController::create",
                "requestBody": {
                    "description": "Route object to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Route"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Route": {
                "description": "A Route defines queue and replyTo queue of a job",
                "properties": {
                    "name": {
                        "description": "The job name",
                        "type": "string"
                    },
                    "queue": {
                        "description": "The queue name",
                        "type": "string"
                    },
                    "replyTo": {
                        "description": "The name of the replyTo queue",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ApiProblem": {
                "description": "Problem Details for HTTP APIs according to RFC7 807",
                "properties": {
                    "type": {
                        "description": " A URI reference [RFC3986] that identifies the problem type.",
                        "type": "string"
                    },
                    "title": {
                        "description": "A short, human-readable summary of the problem type.",
                        "type": "string"
                    },
                    "status": {
                        "description": "The HTTP status code ([RFC7231], Section 6) generated by the origin server for this occurrence of the problem.",
                        "type": "integer"
                    },
                    "detail": {
                        "description": "A human-readable explanation specific to this occurrence of the problem.",
                        "type": "string"
                    },
                    "instance": {
                        "description": "A URI reference that identifies the specific occurrence of the problem.",
                        "type": "string"
                    },
                    "invalidParams": {
                        "type": "array",
                        "items": {
                            "title": "InvalidParameter",
                            "properties": {
                                "name": {
                                    "description": "The name or path of the invalid property",
                                    "type": "string"
                                },
                                "reason": {
                                    "description": "A reason explaining the validation constraint violation",
                                    "type": "string"
                                },
                                "valie": {
                                    "description": "The value that was provided",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
...
    }
}